Plan and Research
Created a plan for the project. I created a rough plan of the project at first with my understanding, researched the project itself and made notes.

Once this was done, I prompted a system prompt for the project (SystemPrompt.txt) to ensure that I was on the right lines, and had not forgotten anything in the plan.

To do this, used Perplexity MCP server inside Cline, to research the tasks at hand, and to create a new plan (including my existing plans), this is plan.md.

Once we had done the research, I updated the system prompt in Cline with (2).

I created StartPrompt.md using a custom GPT, based off the plan.md, to get started. This was entered into Cline, using Claude 3.7 and has System Prompt (2) in the background to guide the AI.